+++ CMakeLists.txt
@@ -24,6 +24,7 @@ add_library(${PROJECT_NAME}_objs OBJECT
     src/api/load_task.cpp
     src/api/get_all_tasks.cpp
     src/api/get_task_by_id.cpp
+    src/api/get_all_solutions.cpp
 )
 target_link_libraries(${PROJECT_NAME}_objs PUBLIC userver::postgresql)
