+++ src/api/get_all_solutions.hpp
@@ -4,7 +4,7 @@
 
 namespace most {
 
-void UnloadAllSolutionsComponent(
+void append_solution_api_handler_get_all_solutions_component(
     userver::components::ComponentList& component_list);
 
 }  // namespace most
