+++ src/main.cpp
@@ -7,6 +7,7 @@
 #include <userver/storages/postgres/component.hpp>
 #include <userver/testsuite/testsuite_support.hpp>
 #include <userver/utils/daemon_run.hpp>
+#include "api/get_all_solutions.hpp"
 #include "api/get_all_tasks.hpp"
 #include "api/get_task_by_id.hpp"
 #include "api/load_task.hpp"
@@ -26,5 +27,8 @@ int main(int argc, char *argv[]) {
     most::append_solution_api_handler_submit_solution_component(component_list);
     most::append_task_api_handler_get_all_tasks_component(component_list);
     most::append_task_api_handler_get_task_by_id_component(component_list);
+    most::append_solution_api_handler_get_all_solutions_component(component_list
+    );
+
     return userver::utils::DaemonMain(argc, argv, component_list);
 }
